refactor: remove pooltool code from chainsync
Provide a notifier API and an example to add it back in the library
consumer code.

Fixes: #20
